Moved metrics() from Node to ScyllaNode
