+++ ccmlib/scylla_node.py
@@ -19,6 +19,7 @@ from six import print_
 from ccmlib import common
 from ccmlib.node import Node
 from ccmlib.node import NodeError
+from ccmlib.prometheus import Prometheus
 
 
 def wait_for(func, timeout, first=0.0, step=1.0, text=None):
@@ -784,3 +785,13 @@ class ScyllaNode(Node):
     def flush(self):
         self.nodetool("flush")
         self._wait_no_pending_flushes()
+
+    def prometheus(self):
+        ip = self.network_interfaces['storage'][0]
+        return Prometheus(ip)
+
+    def metrics(self, select = []):
+        met = self.prometheus().metrics()
+        if select == []:
+            return met
+        return {k:v for k,v in met.iteritems() if k in select}
